python-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
pydantic==2.6.4
grpcio==1.62.1
grpcio-tools==1.62.1
